requirements.txt
lark-parser
llvmlite
colorama
rply